clean up
—move code to external files
—visual styling
—cleanup of stop words and punctuation
